// projects/ui/src/lib/components/po-timeline/enums/po-timeline-mode.enum.ts
/**
 * @usedBy PoTimelineComponent
 *
 * @description
 *
 * Define os tipos modos para o `po-timeline`.
 */
export enum PoTimelineMode {
  /** Itens são exibidos um abaixo do outro intercalando direito e esquerda com 50% de width. */
  Full = 'full',

  /** Itens são exibidos um abaixo do outro com 100% de width. */
  Compact = 'compact'
}